% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSimBzones.R
\name{allocateActivityToSimBzones}
\alias{allocateActivityToSimBzones}
\title{Create SimBzones and allocate total activity among them}
\usage{
allocateActivityToSimBzones(Activity, LocType)
}
\arguments{
\item{Activity}{a number that is the total number of households and jobs
assigned to a location type in an Azone.}

\item{LocType}{a string identifying the location type (Urban, Town, or Rural).}
}
\value{
A numeric vector containing whole number values representing the
amount of activity (households and jobs) in each SimBzone.
}
\description{
\code{allocateActivityToSimBzones} creates a whole number of SimBzones to
accommodate total activity (households and jobs) given a median SimBzone size
for the location type. Allocates activity to the SimBzones.
}
\details{
This function creates a numeric vector containing amounts of activity (jobs
and households). The number of SimBzones is determined by the total activity
and the median SimBzone size for the location type. Total activity is
allocated among the SimBzones in equal proportions using the splitIntegers
function. This function is called by initSimBzones function.
}
