% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignDemandManagement.R
\name{AssignDemandManagement}
\alias{AssignDemandManagement}
\title{Main module function to assign TDM participation and effect on DVMT.}
\usage{
AssignDemandManagement(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{AssignDemandManagement} assigns households and workers to
participation in TDM programs and calculates the proportional reduction in
household DVMT due to this participation.
}
\details{
This function assigns households and workers to participation in TDM programs
and calculates the proportional reduction in household DVMT due to this
participation.
}
