% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSimBzones.R
\name{assignDestAccess}
\alias{assignDestAccess}
\title{Calculate destination accessibility value and level.
\code{assignDestAccess} assign destination accessibility level and value.}
\usage{
assignDestAccess(D1Lvl_Bz, LocType, AreaD5ActProp_D1D5,
  LocTypeD5ActProp_D1D5, D5Ave_D5)
}
\arguments{
\item{D1Lvl_Bz}{A numeric vector of activity density level by SimBzone.}

\item{LocType}{A string identifying the location type}

\item{AreaD5ActProp_D1D5}{A numeric matrix of the destination accessibility
probabilities at each density level for the specific area}

\item{LocTypeD5ActProp_D1D5}{A numeric matrix of the destination accessibility
probabilities at each density level for the location type. Is NULL for Town
and Rural location types}

\item{D5Ave_D5}{A numeric vector identifying the average destination
accessibility at each destination accessibility level}
}
\value{
A list containing 2 components: D5Lvl a character vector identifying
the destination accessibility level of each SimBzone; D5 a numeric vector
containing the destination accessibility level for each SimBzone
}
\description{
This function assigns a destination accessibility level and value to
SimBzones. Destination accessibility is a measure of the proximity of each
SimBzone to population and jobs. This measure is described in detail in the
documentation for the CreateSimBzoneModels.R script. The module assigns a
destination accessibility value and corresponding destination accessibility
level.
}
