% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignParkingRestrictions.R
\name{AssignParkingRestrictions}
\alias{AssignParkingRestrictions}
\title{Main module function to assign parking restrictions to each household.}
\usage{
AssignParkingRestrictions(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{AssignParkingRestrictions} assigns parking restrictions to each
household including residential parking restrictions and worker parking
restrictions.
}
\details{
This function assigns parking restrictions to households. This includes
restrictions on parking household vehicles and restrictions on parking
vehicles at the household workers places of work. The function identifies the
number of free parking spaces the household may use, the charge for household
parking per vehicle if they exceed the number of free spaces, how many
household workers pay for parking, the charge they have to pay, and the
amount of the work parking charges that are 'cash-out-buy-back'.
}
