% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSimBzones.R
\name{calcDensityDistribution}
\alias{calcDensityDistribution}
\title{Determine the activity density (households and jobs per acre) for a set of
SimBzones assigned to a location type in an Azone.}
\usage{
calcDensityDistribution(Activity_Bz, LocType, TargetArea = NULL,
  TargetDensity = NULL, UzaProfileName = NULL)
}
\arguments{
\item{Activity_Bz}{A numeric vector of the amount of activity assigned to
each SimBzone for SimBzones assigned to a location type in an Azone.}

\item{LocType}{A string identifying the location type (Urban, Town, Rural)
that the SimBzones are assigned to.}

\item{TargetArea}{A number identifying the area in acres that is assigned
to accommodate the activity in the location type in the Azone. A value must
be supplied if the LocType is Urban or Town and not supplied if the LocType
is Rural.}

\item{TargetDensity}{A number identifying the average density of activity
(households and jobs per acre) in the LocType in the Azone. A value must be
supplied if the LocType is Rural and not supplied if the LocType is Urban or
Town.}

\item{UzaProfileName}{A string identifying the name of the urbanized area
profile associated with the Marea. A value must be provided if the LocType
is Urban.}
}
\value{
A data frame with columns identifying activity density, activity
density level, and area of each SimBzone.
}
\description{
\code{calcDensityDistribution} calculates the activity density (households
and jobs per acre) for each SimBzone assigned to a location type in an Azone.
}
\details{
This function calculates the activity density (households and jobs per acre)
for each SimBzone assigned to a location type in an Azone. The densities are
calculated so that the overall average density of all the SimBzones equals
the average activity density for the total activity and land area assigned to
the location type. SimBzones are also assign activity density levels which
are used in models to assign destination accessibility and activity diversity.
}
