% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSimBzones.R
\name{calcDiversity}
\alias{calcDiversity}
\title{Assign diversity group and numbers of jobs and households}
\usage{
calcDiversity(Act_Bz, ActDenLvl_Bz, TotEmp, D2ActProp_D1D2, EmpProp_D2_ls,
  LocTyD2ActProp_D1D2)
}
\arguments{
\item{Act_Bz}{A numeric vector of the total activity assigned to each
SimBzone.}

\item{ActDenLvl_Bz}{A character vector of the activity density level assigned
to each SimBzone where activity density is measured as the number of jobs and
households per acre.}

\item{TotEmp}{A number specifying the total number of jobs in the area.}

\item{D2ActProp_D1D2}{A numeric matrix which specifies the proportions of
activity in each diversity group by density group. The rows of the matrix
represent density groups and the columns represent diversity groups.}

\item{EmpProp_D2_ls}{a list having five components, one for each diversity
level. Each component is a list with two components. The 'Values' component
is a numeric vector which specifies values for the employment proportions for
a number of bins. The 'Probs' component is a numeric vector identifying the
probability of each bin.}

\item{LocTyD2ActProp_D1D2}{A numeric matrix which specifies the proportions of
activity in each diversity group by density group for the location type.
For example if the location type is metropolitan, it is the proportions
matrix for all urbanized areas.}
}
\value{
A list having 3 components as follows:
D2Lvl - a character vector identifying the diversity level of each SimBzone,
NumHh - a numeric vector identifying the number of households in each SimBzone
TotEmp - a numeric vector identifying the number of jobs in each SimBzone,
}
\description{
\code{calcDiversity} assigns a diversity group for each SimBzone and splits
activity into numbers of jobs and households.
}
\details{
This function assigns a diversity group to each SimBzone based on the
density profile of the area, the modeled relationship of the
distribution of activity between diversity groups at each density level. The
5 diversity groups are primarily-household, largely-household, mixed,
largely-employment, and primarily-employment. In the case of urbanized areas,
the distribution of activity by diversity group vs. density group
relationship is the modeled relationship from the Smart Location Database.
Average urbanized area values are substituted for density groups missing from
the values for the urbanized area. The proportions are used as sampling
distributions for assigning a diversity group to each SimBzone. The
employment split of activity is modeled using the estimated frequencies of
employment proportions by diversity group. The employment proportions are
adjusted so that the total number of jobs equals the input control total.
}
