% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSimBzones.R
\name{calcNumHhByLocType}
\alias{calcNumHhByLocType}
\title{Calculate number of households by location type}
\usage{
calcNumHhByLocType(NumHh_Az, PropRuralHh_Az, PropTownHh_Az, PropMetroHh_Az,
  Az)
}
\arguments{
\item{NumHh_Az}{A numeric vector of the total number of households in each
Azone.}

\item{PropRuralHh_Az}{A numeric vector identifying the proportion of
households in each Azone that are located in rural locations.}

\item{PropTownHh_Az}{A numeric vector identifying the proportion of
households in each Azone that are located in town locations.}

\item{PropMetroHh_Az}{A numeric vector identifying the proportion of
households in each Azone that are located in metropolitan locations.}

\item{Az}{A character vector of Azone names.}
}
\value{
A list having 3 named components (Rural, Town, Metropolitan) where
each component is a numeric vector identifying the number of households in
the respective location type in each Azone.
}
\description{
\code{calcNumHhByLocType} calculates the number of households by location
type for a set of Azones.
}
\details{
This function calculates the number of households by location type for a
set of Azones as a function of the total number of households by Azone and
user inputs on the proportions of households by location type and Azone.
Location types are metropolitan (i.e. urbanized area), town (i.e. urban areas
that are not urbanized), and rural.
}
