% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSimBzones.R
\name{calcNumJobsByLocType}
\alias{calcNumJobsByLocType}
\title{Calculate number of jobs by location type}
\usage{
calcNumJobsByLocType(NumWkr_Az, PropWkrInRuralJobs_Az,
  PropWkrInTownJobs_Az, PropWkrInMetroJobs_Az, PropMetroJobs_Az, Marea_Az,
  Az)
}
\arguments{
\item{NumWkr_Az}{A numeric vector of the total number of workers residing in
each Azone.}

\item{PropWkrInRuralJobs_Az}{A numeric vector of the proportions of workers
in each Azone that have jobs located in rural locations.}

\item{PropWkrInTownJobs_Az}{A numeric vector of the proportions of workers in
each Azone that have jobs located in town locations.}

\item{PropWkrInMetroJobs_Az}{A numeric vector of the proportions of workers
in each Azone that have jobs located in metropolitan locations.}

\item{PropMetroJobs_Az}{A numeric vector identifying the proportion of
metropolitan jobs for the Marea that the Azone is a part of that are located
in the metropolitan portion of the Azone.}

\item{Marea_Az}{A character vector identifying the Marea associated with each
Azone.}

\item{Az}{A character vector of Azone names.}
}
\value{
A list having 3 named components (Rural, Town, Metropolitan) where
  each component is a numeric vector identifying the number of jobs in the
  respective location type in each Azone.
}
\description{
\code{calcNumJobsByLocType} calculates the number of jobs by location type
for a set of Azones
}
\details{
This function calculates the number of jobs by location type for a set of
Azones as a function of the total number of workers by Azone and user inputs
on the proportions of jobs by location type (metropolitan, town, rural) and
Azone. In addition, the user specifies the proportional allocation of jobs
among the metropolitan portions of Azones that make up an Marea. The function
logic is based on the assumption that Azone workers having jobs in town and
rural locations, will work within the Azone where they reside but that
workers having metropolitan jobs may work in a different Azone portion of the
metropolitan area which includes their Azone.
}
