% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSimBzones.R
\name{calcPlaceTypes}
\alias{calcPlaceTypes}
\title{Assigns an area type and development type to each SimBzone.}
\usage{
calcPlaceTypes(ActDen_Bz, D5_Bz, D2Grp_Bz)
}
\arguments{
\item{ActDen_Bz}{a numeric vector identifying the activity density of each
SimBzone.}

\item{D5_Bz}{a numeric vector identifying the destination accessibility of
each SimBzone.}

\item{D2Grp_Bz}{a character vector identifying the diversity level of each
SimBzone.}
}
\value{
A data frame identifying the area type and development type of each
SimBzone.
}
\description{
\code{calcPlaceTypes} assigns an area type and development type to each
SimBzone.
}
\details{
Place types simplify the characterization of land use patterns. They are used
in the VESimLandUse package modules to simplify the management of inputs for
land use related policies. There are three dimensions to the place type
system. Location type identifies whether the SimBzone is located in an
urbanized area (Metropolitan), a smaller urban-type area (Town), or a
non-urban area (Rural). Area types identify the relative urban nature of the
SimBzone: center, inner, outer, fringe. Development types identify the
character of development in the SimBzone: residential, employment, mix. This
function identifies the area type and development type of each SimBzone as a
function of the activity density, destination accessibility, and diversity
level of the SimBzone.
}
