% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateUrbanMixMeasure.R
\name{CalculateUrbanMixMeasure}
\alias{CalculateUrbanMixMeasure}
\alias{SimulateUrbanMixMeasure}
\title{Main module function that calculates the urban mix measure for each household.}
\usage{
SimulateUrbanMixMeasure(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{SimulateUrbanMixMeasure} calculates the urban mix measure for each
household.
}
\details{
This module calculates whether each household is located in an urban
mixed-use neighborhood based on Bzone density and Bzone input proportion
targets.
}
