% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSimBzones.R
\name{CreateSimBzones}
\alias{CreateSimBzones}
\title{Create SimBzones.}
\usage{
CreateSimBzones(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{CreateSimBzones} creates SimBzones assigned to each Azone to
accommodate projected households and jobs in each Azone.
}
\details{
This function creates SimBzones assigned to each Azone to accommodate
projected households and jobs in each Azone and to reflect likely
distribution of characteristics for activity density, mixing of jobs and
households, and destination accessibility. The module also identifies the
location type, area type, and development type of each SimBzone.
}
