% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSimBzones.R
\docType{data}
\name{CreateSimBzonesSpecifications}
\alias{CreateSimBzonesSpecifications}
\title{Specifications list for CreateSimBzones module}
\format{A list containing 5 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{NewSetTable}{new table to be created for datasets specified in the
 'Set' specifications}
 \item{Inp}{scenario input data to be loaded into the datastore for this
 module}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\source{
CreateSimBzones.R script.
}
\usage{
CreateSimBzonesSpecifications
}
\description{
A list containing specifications for the CreateSimBzones module.
}
\keyword{datasets}
