% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateUrbanMixMeasure.R
\name{idUrbanMixBzones}
\alias{idUrbanMixBzones}
\title{Identify urban mix Bzones}
\usage{
idUrbanMixBzones(Hh_BzHt, Den_Bz, Target)
}
\arguments{
\item{Hh_BzHt}{a numeric matrix of the number of households by Bzone and by
housing type. The rownames of the matrix are the Bzone names. The column
names are the housing type names: SFDU (single family dwelling unit), MFDU
(multifamily dwelling unit)}

\item{Den_Bz}{a named numeric vector of the population density (persons per
square mile) of each Bzone. The names of the vector are the Bzone names.}

\item{Target}{a target for the proportion of households in the Marea who live
in urban mixed neighborhoods.}
}
\value{
a named numeric vector which has a value of 1 for Bzones that are
are identified as being urban mix and 0 for Bzones that are not. The names of
the vector are the Bzone names.
}
\description{
\code{idUrbanMixBzones} identifies Bzones that are urban mix and Bzones
that are not.
}
\details{
This function identifies which Bzones in an Marea are urban mixed use and
those that are not as a function of the Bzone population density and the mix
of single family and multifamily dwellings. The UrbanMixModel estimated by
the CalculateUrbanMixMeasure module in the VELand use package is used to
estimate the probability for single family households and the probability for
multifamily dwellings in each Bzone. A weighted average probability is
calculated for each Bzone where the numbers of single family and multifamily
dwellings are the weights. The descending order of probability is determined
and the minimum number of zones in that order necessary to meet the target is
determined. Those zones are classified as urban mixed use. If a numerical
target is not specified, the function calculates an expected proportion of
households using the zonal weighted probabilities.
}
