% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSimBzones.R
\name{initSimBzones}
\alias{initSimBzones}
\title{Creates SimBzones to accommodate the activity assigned to a set of Azones.}
\usage{
initSimBzones(RuralActivity_Az, TownActivity_Az, UrbanActivity_Az, Az,
  Marea_Az)
}
\arguments{
\item{RuralActivity_Az}{a numeric vector identifying the amount of activity
assigned to the rural location type in each Azone.}

\item{TownActivity_Az}{a numeric vector identifying the amount of activity
assigned to the town location type in each Azone.}

\item{UrbanActivity_Az}{a numeric vector identifying the amount of activity
assigned to the urban location type in each Azone.}

\item{Az}{a character vector identifying the name of each Azone.}

\item{Marea_Az}{a character vector identifying the name of the Marea that
each Azone is associated with.}
}
\value{
A data frame containing Bzone names, Azone names, Marea names,
LocType, and activity, where each row represents a SimBzone.
}
\description{
\code{initSimBzones} creates a set of SimBzones which accomodates the
activity assigned to location types in Azones.
}
\details{
This function creates a set of SimBzones for a set of Azones which
accommodates the activity (households and jobs) assigned to each location
type (Urban, Town, Rural) in each of the Azones. It calls the
allocateActivity function to create a vector of SimBzones activity amounts to
accommodate the total activity assigned to to a location type in an Azone. It
creates a data frame which in addition to containing the activity allocation,
assigns a unique ID to each SimBzone and also identifies the Azone and Marea
the SimBzone is in. It also identifies the location type (LocType) of each
SimBzone.
}
