% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSimBzoneModels.R
\docType{data}
\name{SimBzone_ls}
\alias{SimBzone_ls}
\title{SimBzone model parameters}
\format{A list having the following components:
\describe{
  \item{UaProfiles}{a list containing profiles for urbanized areas}
  \item{TnProfiles}{a list containing profiles for towns}
  \item{RuProfiles}{a list containing profiles for rural areas}
  \item{Abbr}{a list containing dimension naming vectors}
  \item{Docs}{a list containing miscellaneous model documentation objects}
  \item{EmpSplit}{a list containing two matrices of employment proportions by placetype and quantile}
  \item{HousingSplit}{a list containing a matrix of multi-family dwelling unit proportions by placetype and quantile}
}}
\source{
CreateSimBzoneModels.R script.
}
\usage{
SimBzone_ls
}
\description{
A list containing various parameters for SimBzone models
}
\keyword{datasets}
