% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate4DMeasures.R
\name{Simulate4DMeasures}
\alias{Simulate4DMeasures}
\title{Main module function that simulates 4D measures for each Bzone.}
\usage{
Simulate4DMeasures(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{Simulate4DMeasures} simulates 4D measures for each Bzone.
}
\details{
This module simulates several 4D measures by Bzone including density,
diversity (i.e. mixing of land uses), design (i.e. multimodal network design),
and destination accessibility.
}
