% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate4DMeasures.R
\docType{data}
\name{Simulate4DMeasuresSpecifications}
\alias{Simulate4DMeasuresSpecifications}
\title{Specifications list for Simulate4DMeasures module}
\format{A list containing 4 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Inp}{scenario input data to be loaded into the datastore for this
 module}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\source{
Simulate4DMeasures.R script.
}
\usage{
Simulate4DMeasuresSpecifications
}
\description{
A list containing specifications for the Simulate4DMeasures module.
}
\keyword{datasets}
