% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate4DMeasures.R
\name{simulateD3bpo4}
\alias{simulateD3bpo4}
\title{Function which simulates D3bpo4 values for SimBzones}
\usage{
simulateD3bpo4(AveD3bpo4_, PlaceType_, NormD3bpo4_PtQt)
}
\arguments{
\item{AveD3bpo4_}{a numeric vector identifying for all the SimBzones_ the
average D3bpo4 value for the place where each SimBzone is located. For
urban locations it is the urbanized area. For town and rural places it is the
average value for towns and rural places respectively.}

\item{PlaceType_}{a character vector identifying for all the SimBzones_ the
place type for each SimBzone. Place type is created by concatenating the
AreaType and DevType designations using a period (.) separator.}

\item{NormD3bpo4_PtQt}{a numeric matrix which tabulates normalized D3bpo4
values by place type and quantile. See the CreateSimBzoneModels
documentation.}
}
\value{
A numeric vector containing D3bpo4 values for the SimBzones.
}
\description{
\code{simulateD3bpo4} simulates D3bpo4 values for SimBzones
}
\details{
This function simulates pedestrian network design measure, D3bpo4, as defined
by the U.S. Environmental Protection Agency (EPA) in the Smart Location
Database (SLD). SLD documentation is included in the inst/extdata/sources
directory of the VESimLandUseData package. In short, D3bpo4 is a measure of
'intersection density in terms of pedestrian-oriented intersections having
four or more legs per square mile'. D3bpo4 is simulated as a function of the
place type of the SimBzone (LocType, AreaType, DevType) and the average
D3bpo4 value for the urbanized area (or town or rural area) where the
SimBzone is located. A base average value by urbanized area and towns as a
whole and rural areas as a whole in the CreateSimBzoneModels module. User
inputs adjust the values (e.g. increase average by 50%). The D3bpo4
model estimated by the CreateSimBzoneModels module is used to calculate the
SimBzone value.
}
