% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateEmployment.R
\name{SimulateEmployment}
\alias{SimulateEmployment}
\title{Main module function to assign employment by type to Bzones.}
\usage{
SimulateEmployment(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{SimulateEmployment} assigns workers to Bzones.
}
\details{
This function assigns workers to Bzone employment based on inputs of
employment by type by Bzone, adjustment of employment to equal regionwide
total of workers, and identify Bzone employment location for each worker as
a function of the number of jobs in each Bzone and the inverse of distance
between Bzones.
}
