% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateHousing.R
\name{SimulateHousing}
\alias{SimulateHousing}
\title{Main module function that assigns the housing type and Bzone for each
household.}
\usage{
SimulateHousing(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{SimulateHousing} predicts the housing type and Bzone for each household.
}
\details{
This function predicts the housing choice of each household. It uses the
estimated models of binomial choice model for determining the probability
that the housing choice for each household is single family (SF) vs.
multifamily (MF). The group quarters population is assigned to group quarters
(GQ). After the housing choice is assigned, the household is assigned to a
Bzone based on the input assumptions of:
1) housing supply by type for each Bzone
2) distribution of households by income quartile for each Bzone
Housing demand by Bzone, type, and income quartile is balanced allocated
by iterative proportional fitting. Households are assigned randomly to Bzones
based on their housing choice and income quartile and the balanced quantity
of housing in each Bzone by income quartile and housing type.
}
