% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSimBzones.R
\name{splitEmployment}
\alias{splitEmployment}
\title{Split SimBzone employment into retail, service, and other components.}
\usage{
splitEmployment(TotEmp_, PlaceType_, RetSvcProp_PtQt, RetProp_PtQt)
}
\arguments{
\item{TotEmp_}{A numeric vector identifying the total number of jobs assigned
to each SimBzone.}

\item{PlaceType_}{A character vector identifying the place type of each
SimBzone where the place type designation is a concatenation of area type and
development type joined by a period ('.').}

\item{RetSvcProp_PtQt}{A matrix of the quantiles of retail and service
employment proportion of total employment by place type and quantile. See the
documentation for the CreateSimBzoneModels.R script for more information.}

\item{RetProp_PtQt}{A matrix of the quantiles of retail employment proportion
of retail and service employment by place type and quantile. See the
documentation for the CreateSimBzoneModels.R script for more information.}
}
\value{
A data frame identifying the numbers of retail, service, and other
jobs located in each SimBzone.
}
\description{
\code{splitEmployment} splits the total employment of each SimBzone into
retail, service, and other components.
}
\details{
This function splits the employment of each SimBzone in a set of SimBzones
into retail, service, and other components.
}
