% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSimBzones.R
\name{splitHousing}
\alias{splitHousing}
\title{Split SimBzone housing into single family and multifamily.}
\usage{
splitHousing(Hh_, PlaceType_, MFProp_PtQt)
}
\arguments{
\item{Hh_}{a numeric vector identifying the total number of households in
each SimBzones}

\item{PlaceType_}{a character vector identifying the place type of each
SimBzone where place type is the concatenation of AreaType and DevType with
a period '.' separator.}

\item{MFProp_PtQt}{a numeric matrix of multifamily housing proportions by
place type and quantile created by the 'CreateSimBzoneModel' module.}
}
\value{
A data frame identifying the numbers of single family dwelling units
and multifamily dwelling units in each SimBzone.
}
\description{
\code{splitHousing} splits the total number of households in each SimBzone
into single family and multifamily components.
}
\details{
This function splits the total number of households in each SimBzone into
single family and multifamily components using the housing split model
estimated and documented in the 'CreateSimBzoneModel' module.
}
