% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateUrbanMixMeasure.R
\docType{data}
\name{UrbanMixModel_ls}
\alias{UrbanMixModel_ls}
\title{Urban mixed-use model}
\format{A list having the following components:
\describe{
  \item{Type}{a string identifying the type of model ("binomial")}
  \item{Formula}{makeModelFormulaString(UrbanMixModel)}
  \item{PrepFun}{a function that prepares inputs to be applied in the model}
  \item{Summary}{the summary of the binomial logit model estimation results}
  \item{SearchRange}{a two-element vector specifying the range of search values}
}}
\source{
CalculateUrbanMixMeasure.R script.
}
\usage{
UrbanMixModel_ls
}
\description{
A list containing the model equation and other information needed to
implement the urban mixed-use model.
}
\keyword{datasets}
