% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSimBzoneModels.R
\docType{data}
\name{UsaBlkGrpTypes_df}
\alias{UsaBlkGrpTypes_df}
\title{Measures of land use types and levels for block groups}
\format{A data frame having the following components:
\describe{
  \item{GEOID10}{a string that identifies the unique block group ID}
  \item{LocType}{a string that identifies the location type}
  \item{D1DGrp}{a factor that identifies the activity density level}
  \item{D2Grp}{a factor that identifies the land use mixing level}
  \item{D5Grp}{a factor that identifies the destination accessibility level}
  \item{AreaType}{a string that identifies the area type}
  \item{DevType}{a string that identifies the development type}
}}
\source{
CreateSimBzoneModels.R script.
}
\usage{
UsaBlkGrpTypes_df
}
\description{
A data frame containing values of for land use types and levels for land use
density, mixing, and destination accessibility measures corresponding to
block groups in the VESimLandUseData::SimLandUseData_df data frame
}
\keyword{datasets}
