% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialize.R
\docType{data}
\name{UzaProfileNames_ls}
\alias{UzaProfileNames_ls}
\title{Urbanized area profile names list}
\format{A list containing 2 components:
\describe{
 \item{Names}{a sorted vector of urbanized area names}
 \item{Inp}{a 3-column data frame of the names to include in documentation}
}}
\source{
Initialize.R script.
}
\usage{
UzaProfileNames_ls
}
\description{
A list containing the names of urbanized areas included in the profiles of
urbanized area land use characteristics.
}
\keyword{datasets}
